proyectoxEtapas/etapa5-python-flask/2_intermedio/otro/introWebModeloOSIEnClase/requirements.txt
# Archivo de dependencias para la aplicación web de redes de computadoras
# Este archivo lista todas las librerías de Python necesarias para el proyecto

# Framework web principal
Flask==3.0.0

# Conector de MySQL para Python
mysql-connector-python==8.2.0

# Librerías adicionales que pueden ser útiles
Werkzeug==3.0.1
Jinja2==3.1.2
MarkupSafe==2.1.3
itsdangerous==2.1.2
click==8.1.7
blinker==1.7.0

# Para desarrollo (opcional)
# python-dotenv==1.0.0  # Para variables de entorno
# flask-cors==4.0.0     # Para CORS si es necesario

# NOTA: Bootstrap 5.3.2 está incluido localmente en static/css/bootstrap.min.css
# y static/js/bootstrap.bundle.min.js - no requiere instalación via pip 